# CUDA: add bf16 and f32 support to cublas_mul_mat_batched