Add bf16 and f32 to support batched cuBLAS mul mat. Speed up when we do 

`--cache_type_v bf16 --cache_type_k bf16` when running llama-bench

| Model           | Test   |   t/s master |   t/s add_bp16_fp32_to_cublas_batched |   Speedup |
|:----------------|:-------|-------------:|--------------------------------------:|----------:|
| llama 7B Q5_K_M | pp512  |      3681.87 |                               4807.34 |      1.31 |
| llama 7B Q5_K_M | tg128  |       129.40 |                                129.28 |      1.00 |