@am17an 2025-06-25
mul_mat_batched with bf16 is failing for `ubuntu-22-cmake-vulkan`, should I remove the extra tests?