@am17an 2025-06-25
> mul_mat_batched with bf16 is failing for `ubuntu-22-cmake-vulkan`, should I remove the extra tests?

@JohannesGaessler mul-mat tests in bf16 which fail for vulkan because of an assert `/home/runner/work/llama.cpp/llama.cpp/ggml/src/ggml-vulkan/ggml-vulkan.cpp:5134: GGML_ASSERT(ggml_vk_dim01_contiguous(src0) || src0->type == GGML_TYPE_F32 || src0->type == GGML_TYPE_F16) failed` . Not familiar with the vulkan code