@JohannesGaessler 2025-06-25
Sorry, I didn't see the Vulkan comment. The problem from what I can tell is that the logic in `ggml_backend_vk_device_supports_op` and the assert are inconsistent. What I assume happened is that one of the 2 was changed without updating the other one, so presumably the problem can be fixed by updating the outdated one (and until now this simply wasn't noticed). @0cc4m @jeffbolznv can either of you weigh in on which version is correct?