@jeffbolznv 2025-06-25
I think this was supposed to work, but just changing the assert I see the test fail. I'll debug it.