@jeffbolznv 2025-06-25
https://github.com/ggml-org/llama.cpp/pull/14378 should fix the new tests.